add correct for homework 4 to git
