+++ src/test/java/tests/LoginTest.java
@@ -7,8 +7,6 @@ import org.testng.annotations.Test;
 
 public class LoginTest extends BaseTest {
 
-    private static final Logger log = LoggerFactory.getLogger(LoginTest.class);
-
     @Test
     public void checkCorrectLogin() {
         loginPage.open();
@@ -45,32 +43,4 @@ public class LoginTest extends BaseTest {
                 "Epic sadface: Username is required",
                 "Сообщение об ошибке не появилось");
     }
-
-    @Test
-    public void checkAddProductToCart() {
-        loginPage.open();
-        loginPage.login("standard_user", "secret_sauce");
-        Assert.assertEquals(productsPage.getTitle(),
-                "Products",
-                "Логин не выполнен");
-        productsPage.clickButton();
-        Assert.assertEquals(productsPage.checkButtonRemove(),
-                "Remove",
-                "Кнопка Удалить не появилась");
-        cartPage.clickToCartLink();
-        //проверяем, что мы на странице с title = You cart
-        Assert.assertEquals(cartPage.getTitle(),
-                "Your Cart",
-                "Переход в корзину не удался");
-        //проверить, что кнопка удалить именно от добавленного товара присутствует на странице корзины
-        Assert.assertEquals(productsPage.checkButtonRemove(),
-                "Remove",
-                "Кнопка Удалить не появилась");
-        inventoryPage.clickToInventoryLink();
-        // проверяем, что открыта страница товара
-        Assert.assertEquals(inventoryPage.getNameProduct(),
-                "Sauce Labs Backpack",
-                "Открыт не тот товар");
-    }
-
 }
